bz18603: Fixes for removing files from metadata manager
Remove paths from the db cache on remove() instead of removed_from_db().
Waiting for removed_from_db() doesn't work well if we're inside a bulk sql
operation.

Added unittests.

+++ tv/lib/item.py
@@ -2187,9 +2187,17 @@ class FileItem(Item):
                 self.make_deleted()
         if old_parent is not None and old_parent.get_children().count() == 0:
             old_parent.expire()
-        app.local_metadata_manager.remove_file(self.filename)
+        if app.local_metadata_manager.path_in_system(self.filename):
+            app.local_metadata_manager.remove_file(self.filename)
+
+    def remove(self):
+        if app.local_metadata_manager.path_in_system(self.filename):
+            app.local_metadata_manager.remove_file(self.filename)
+        Item.remove(self)
 
     def make_deleted(self):
+        if app.local_metadata_manager.path_in_system(self.filename):
+            app.local_metadata_manager.remove_file(self.filename)
         self._remove_from_playlists()
         self.downloadedTime = None
         # Move to the manual feed, since from Miro's point of view the file is
@@ -2202,6 +2210,7 @@ class FileItem(Item):
     def make_undeleted(self):
         self.deleted = False
         self.signal_change()
+        app.local_metadata_manager.add_file(self.filename)
 
     def delete_files(self):
         if self.has_parent():
